PS-9089: [AA CR] Reverted previous ShoppingListEntityManager changes, Added shopping list item existence check inside shopping list item deletion transaction

+++ Bundles/ShoppingList/src/Spryker/Zed/ShoppingList/Business/Model/ShoppingListItemOperation.php
@@ -329,10 +329,20 @@ class ShoppingListItemOperation implements ShoppingListItemOperationInterface
      */
     protected function deleteShoppingListItemTransaction(ShoppingListItemTransfer $shoppingListItemTransfer): ShoppingListItemResponseTransfer
     {
+        $shoppingListItemCollectionTransfer = $this->shoppingListRepository->findShoppingListItemsByIds([
+            $shoppingListItemTransfer->getIdShoppingListItem(),
+        ]);
+
+        if (!$shoppingListItemCollectionTransfer->getItems()->count()) {
+            return (new ShoppingListItemResponseTransfer())
+                ->setIsSuccess(false);
+        }
+
         $this->pluginExecutor->executeBeforeDeletePlugins($shoppingListItemTransfer);
         $this->shoppingListEntityManager->deleteShoppingListItem($shoppingListItemTransfer->getIdShoppingListItem());
 
-        return (new ShoppingListItemResponseTransfer())->setIsSuccess(true);
+        return (new ShoppingListItemResponseTransfer())
+            ->setIsSuccess(true);
     }
 
     /**
